+++ certbot-dns-dnsimple/certbot_dns_dnsimple/_internal/__init__.py
@@ -1,0 +1,1 @@
+Internal implementation of `~certbot_dns_dnsimple.dns_dnsimple` plugin.
