+++ certbot-dns-dnsimple/docs/api/dns_dnsimple.rst
@@ -1,5 +1,5 @@
-:mod:`certbot_dns_dnsimple.dns_dnsimple`
+:mod:`certbot_dns_dnsimple._internal.dns_dnsimple`
 ----------------------------------------
 
-.. automodule:: certbot_dns_dnsimple.dns_dnsimple
+.. automodule:: certbot_dns_dnsimple._internal.dns_dnsimple
    :members:
